+++ CHANGELOG.md
@@ -21,6 +21,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - **CUMULUS-2814**
     - Remove DynamoDB logic from rules `DELETE` endpoint
     - Move event resources deletion logic from `rulesModel` to `rulesHelper`
+  - **CUMULUS-2815**
+    - Remove update of DynamoDB record from API endpoint PUT /collections/<name>/<version>
+    - Move File Config and Core Config validation logic for Postgres Collections from `api/models/collections.js` to `api/lib/utils.js`   
   - **CUMULUS-2813**
     - Removes creation and deletion of DynamoDB record from API endpoint POST /rules/